requirements/testing.txt
-r development.txt

coverage==6.3.2
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2